:recycle: Switch LogService to Sequelized version.
Also tweak to provide a more standard interface.
This probably needs a lot more work and having the table rebuild.
The Log should probably use a external library and be a polymorphic table.
e.g.
entity_id
entity_type
and the operation should probably be a single word like the LogOperationTypes enum.