:beetle: Fix DataTypes.NOW bug for MSSQL, now sets default as `GETDATE()`
Note that in the models, I'm still using `DataTypes.NOW`.
I'm not sure if this is correct or not.
I _think_ DataTypes.NOW sets the value as new Date() at the JS level instead of the database so its still useful.

Also add a warning once we upgrade to Sequelize 7 that we probably don't need this kind of shim anymore.