:beetle: Fix formatting lost on second render of component.
e.g. toggling between tabs on the Employees tab of the child care centre details view.
http://localhost:8080/child-care-centres/1/2023-24/employees/april to http://localhost:8080/child-care-centres/1/2023-24/employees/may and then back resulted in 5000 instead of $5,000.00

This also, incidentally fixes the jump to end cursor bug. I'm not really sure how.
Possibly because I'm no longer passing the value to the v-text field? And am instead simply relying on the library to do that for me? I guess the library just knows to load/inject this.modelValue?