:japanese_castle: Force invalid api calls to bounce the user to the login page.
Why? There have been reports that payments aren't persisting after logout, and since I haven't been able to recreate the issue locally, the only other thing I can think of is that the user's session had silently expired.

Also this is generally a good practice, though it might be better to handle this in a route guard or global watcher or something.