:recycle: Remove db service health check in favor of JS level health check.
Why? MSSQL released a new version that broke the docker level health check.
This lead me to reflect on whether the docker level health check was necessary.
I concluded that it was both hard to read and debug, and that it was not necessary.

Use code patterns from https://github.com/icefoganalytics/wrap/pull/86