:tophat: Update entity relationship diagrams.
NOTE: centres funding_region_id was supposed to be nullable, but it was not.
The migration api/src/db/migrations/2025.12.09T15.12.00.add-funding-region-id-to-centres.ts
failed to change the funding_region_id column to nullable due to a bug in Sequelize with MSSQL.

We should switch to Knex as the effects are more predicatable. See https://yg-hpw.atlassian.net/browse/ELCC-87